requirements.txt
gitpython==3.1.40
google-generativeai>=0.8.0
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.0